% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrWH.R
\name{distrWH}
\alias{distrWH}
\title{Distribution of records}
\usage{
distrWH(x, dec = FALSE)
}
\arguments{
\item{x}{Numeric. A vector containing the data set.}

\item{dec}{Numeric. A value defined to set the number of decimals in the returns.}
}
\value{
List including the table with the absolute frequency of data, also the mean, median, mode and range of data.
}
\description{
This function is useful to see the frequency of data and descriptive statistics
}
\details{
This function shows statistical descriptive values of a vector.
}
\examples{
#Example 1
dat=c(1,5,2,4,5,1,3,5,3,5)
distrWH(x=dat,dec=3)
}
